requirements.txt
certifi==2019.6.16
chardet==3.0.4
html-testRunner==1.2.1
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
pathlib==1.0.1
pyunitreport==0.1.4
requests==2.22.0
selenium==3.141.0
urllib3==1.25.3
